Group-Service Implementationen und Hülle für Message

+++ server/src/main/java/service/classes/GroupServiceImpl.java
@@ -6,7 +6,10 @@ import dao.interfaces.GroupDao;
 import dao.interfaces.Group_UserDao;
 import model.interfaces.Group;
 import model.interfaces.User;
+import service.exceptions.GroupDoesNotExistException;
 import service.interfaces.GroupService;
+
+import static service.constants.ServiceConstants.ERR_MSG_GROUP_DOES_NOT_EXIST;
 
 import java.util.ArrayList;
 
@@ -25,53 +28,91 @@ public class GroupServiceImpl implements GroupService {
         groupDao.addNewGroup(group);
     }
 
-    public void addUserToGroup(int id, String username) {
-        //TODO: Implentierung
-    }
-
-    public void addUserToGroup(int userId, int id) {
-        //TODO: Implentierung
+    public void addUserToGroup(int id, int userId) {
+        User user = ModelObjectBuilder.getUserObject();
+        user.setUserId(userId);
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        Group group = ModelObjectBuilder.getGroupObject();
+        group.setGroupId(id);
+        group_userDao.addNewUser(user, group);
     }
 
     public void addUserToGroup(int id, User user) {
-        //TODO: Implentierung
+        Group group = ModelObjectBuilder.getGroupObject();
+        group.setGroupId(id);
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        group_userDao.addNewUser(user, group);
     }
 
-    public void removeUserFromGroup(int id, String username) {
-        //TODO: Implentierung
-    }
     public void removeUserFromGroup(int id, int userId) {
-        //TODO: Implentierung
+        User user = ModelObjectBuilder.getUserObject();
+        user.setUserId(userId);
+        Group group = ModelObjectBuilder.getGroupObject();
+        group.setGroupId(id);
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        group_userDao.removeUser(user, group);
     }
 
     public void removeGroup(int id) {
-        //TODO: GroupDoesNotExistException
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
         groupDao.removeGroup(getGroupDao(id));
     }
 
     public Group getGroupById(int id) {
-        //TODO: GroupDoesNotExistException
-        return groupDao.getGroupById(getGroupDao(id));
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        return getGroupDao(id);
     }
 
     public void changeGroupName(int id, String newName) {
-        //TODO: GroupDoesNotExistException
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
         Group group = getGroupDao(id);
         group.setGroupName(newName);
         groupDao.changeGroupName(group);
     }
 
    public ArrayList<User> getUsersForGroup(int id) {
-       //TODO: Implentierung
-        return null;
+       if(!doesGroupExist(id)) {
+           throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+       }
+        return group_userDao.getUsersByGroup(getGroupDao(id));
     }
     public ArrayList<Integer> getUserIdsForGroup(int id) {
-        //TODO: Implentierung
-        return null;
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        ArrayList<User> usersInGroup = getUsersForGroup(id);
+        ArrayList<Integer> userIds = null;
+        for (User u : usersInGroup) {
+            userIds.add(u.getUserId());
+        }
+        return userIds;
     }
     public ArrayList<String> getUserNamesForGroup(int id) {
-        //TODO: Implentierung
-        return null;
+        if(!doesGroupExist(id)) {
+            throw new GroupDoesNotExistException(ERR_MSG_GROUP_DOES_NOT_EXIST);
+        }
+        ArrayList<User> usersInGroup = getUsersForGroup(id);
+        ArrayList<String> userNames = null;
+        for (User u : usersInGroup) {
+            userNames.add(u.getUsername());
+        }
+        return userNames;
+    }
+
+    public boolean doesGroupExist(int id) {
+        return getGroupDao(id).equals(ModelObjectBuilder.getGroupObject());
     }
 
     private Group getGroupDao(int id) {
@@ -79,4 +120,5 @@ public class GroupServiceImpl implements GroupService {
         group.setGroupId(id);
         return groupDao.getGroupById(group);
     }
+
 }
